Buffer overflow and timeout
SCPI_Parser::GetMessage
A message buffer overflow check is added.
If an overflow occours, we try to read until the next termination chars
and discard that message, also "buffer_overflow" is set to true.

The fixed 10ms timeout time is changed by the SCPI_TIMEOUT macro.
If a timeout event occours, then "timeout" is set to true.

Both "buffer_overflow" and "timeout" are set to false on the next
SCPI_Parser::GetMessage call.